Move load and open out of __init__.py
